+++ truck-assembly/src/dag.rs
@@ -826,15 +826,20 @@ impl<NE, EE> Dag<NE, EE> {
     /// assert_eq!(edges1[0].nodes(), (b[1].index(), b[2].index()));
     /// assert_eq!(*edges1[0].entity(), 9.0);
     /// ```
-    pub fn map<NE2, EE2, NF, EF>(&self, mut node_map: NF, mut edge_map: EF) -> Dag<NE2, EE2>
+    pub fn map<'a, NE2, EE2, NF, EF>(
+        &'a self,
+        mut node_map: NF,
+        mut edge_map: EF,
+    ) -> Dag<NE2, EE2>
     where
-        NF: FnMut(&NE) -> NE2,
-        EF: FnMut(&EE) -> EE2, {
-        let mut edge_data_map = move |edge_data: &EdgeData<EE>| EdgeData {
+        NF: FnMut(&'a NE) -> NE2,
+        EF: FnMut(&'a EE) -> EE2,
+    {
+        let mut edge_data_map = move |edge_data: &'a EdgeData<EE>| EdgeData {
             to: edge_data.to,
             entity: edge_map(&edge_data.entity),
         };
-        let node_data_map = move |node_data: &NodeData<NE, EE>| NodeData {
+        let node_data_map = move |node_data: &'a NodeData<NE, EE>| NodeData {
             entity: node_map(&node_data.entity),
             edges: node_data.edges.iter().map(&mut edge_data_map).collect(),
             parents: node_data.parents.clone(),
@@ -879,19 +884,19 @@ impl<NE, EE> Dag<NE, EE> {
     /// assert_eq!(edges1[0].nodes(), (b[1].index(), b[2].index()));
     /// assert_eq!(*edges1[0].entity(), 9.0);
     /// ```
-    pub fn par_map<NE2, EE2, NF, EF>(&self, node_map: NF, edge_map: EF) -> Dag<NE2, EE2>
+    pub fn par_map<'a, NE2, EE2, NF, EF>(&'a self, node_map: NF, edge_map: EF) -> Dag<NE2, EE2>
     where
         NE: Sync,
         EE: Sync,
         NE2: Send,
         EE2: Send,
-        NF: Fn(&NE) -> NE2 + Send + Sync,
-        EF: Fn(&EE) -> EE2 + Send + Sync, {
-        let edge_data_map = move |edge_data: &EdgeData<EE>| EdgeData {
+        NF: Fn(&'a NE) -> NE2 + Send + Sync,
+        EF: Fn(&'a EE) -> EE2 + Send + Sync, {
+        let edge_data_map = move |edge_data: &'a EdgeData<EE>| EdgeData {
             to: edge_data.to,
             entity: edge_map(&edge_data.entity),
         };
-        let node_data_map = move |node_data: &NodeData<NE, EE>| NodeData {
+        let node_data_map = move |node_data: &'a NodeData<NE, EE>| NodeData {
             entity: node_map(&node_data.entity),
             edges: node_data.edges.iter().map(&edge_data_map).collect(),
             parents: node_data.parents.clone(),