+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ## Unreleased
 
+- Fix syntax error when reading binary STL > 8192b.
+- Fillet for a single edge.
+- Write assembly to step file.
 - In the README, we clarified that the subtitle is the origin of the name “truck,” and changed all instances of the term in the main text to `truck`.
 - Get more precise part attributions from `Product` and `NextAssemblyUsageOccurrence`.
 - Add the variable `division` to `truck_modeling::builder::rsweep`.
@@ -62,7 +65,7 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ### Latest `cargo upgrade`
 
-2026-02-06
+2026-03-13
 
 ## v0.6